% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_cdr_jobs.R
\name{calculate_cdr_jobs}
\alias{calculate_cdr_jobs}
\title{Calculate CDR Jobs}
\usage{
calculate_cdr_jobs(
  db_path,
  db_name,
  dat_file = NULL,
  scenario_list,
  region_list,
  output_path,
  output_type = "csv",
  create_plots = TRUE,
  job_metric = "max_Jobs",
  ncol = 2,
  nrow = NULL,
  selected_years = NULL
)
}
\description{
Calculate CDR Jobs
}
